开始开发接口了

+++ src/main/java/com/yunhuakeji/attendance/enums/ClockStatus.java
@@ -4,20 +4,20 @@ package com.yunhuakeji.attendance.enums;
  * 考勤状态枚举
  */
 public enum ClockStatus {
-    NOT_CLOCK(1, "未打卡"),
-    CLOCK(2, "到勤"),
-    STAYOUT_LATE(3, "晚归"),
-    STAYOUT(4, "未归");
+    NOT_CLOCK((byte) 1, "未打卡"),
+    CLOCK((byte) 2, "到勤"),
+    STAYOUT_LATE((byte) 3, "晚归"),
+    STAYOUT((byte) 4, "未归");
 
-    private int type;
+    private byte type;
     private String msg;
 
-    ClockStatus(int type, String msg) {
+    ClockStatus(byte type, String msg) {
         this.type = type;
         this.msg = msg;
     }
 
-    public static ClockStatus get(int val) {
+    public static ClockStatus get(byte val) {
         for (ClockStatus c : ClockStatus.values()) {
             if (c.getType() == val) {
                 return c;
@@ -26,7 +26,7 @@ public enum ClockStatus {
         return null;
     }
 
-    public int getType() {
+    public byte getType() {
         return type;
     }
